cosmocc: new port
Adds Cosmopolitan Libc 4.0.2 and its compiler toolchain to MacPorts.
The patched package.sh script has been updated upstream to support
building on MacOS, but is included here as a backport.

A launcher is included to add the cosmocc toolchain to MacPorts bin/.
Due to cosmocc's directory structure, it is better off isolated in
its own subdirectory, which also avoids polluting the MacPorts include/
directory with Cosmopolitan headers.